requirements.txt
discord.py==1.7.3
requests==2.26.0
discord-components==2.0.4
fuzzywuzzy==0.18.0
attrs==21.2.0
python-Levenshtein==0.12.2
dnspython==2.1.0
pymongo==3.12.0
aiohttp==3.7.4.post0
aiodns==3.0.0
Pillow==8.3.1
asyncpg==0.24.0
onnxruntime==1.8.1
opencv-python-headless==4.5.3.56